+++ cmd/neofs-node/accounting.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"github.com/nspcc-dev/neofs-node/pkg/morph/client/balance"
 	accountingService "github.com/nspcc-dev/neofs-node/pkg/services/accounting"
 	protoaccounting "github.com/nspcc-dev/neofs-sdk-go/proto/accounting"
 )
@@ -11,10 +10,7 @@ func initAccountingService(c *cfg) {
 		initMorphComponents(c)
 	}
 
-	balanceMorphWrapper, err := balance.NewFromMorph(c.cfgMorph.client, c.balanceSH)
-	fatalOnErr(err)
-
-	server := accountingService.New(&c.key.PrivateKey, c.networkState, balanceMorphWrapper)
+	server := accountingService.New(&c.key.PrivateKey, c.networkState, c.bCli)
 
 	for _, srv := range c.cfgGRPC.servers {
 		protoaccounting.RegisterAccountingServiceServer(srv, server)